% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_BIC_MDS_list.R
\name{compute_BIC_MDS.list}
\alias{compute_BIC_MDS.list}
\title{Compute BIC for MDS according to Lee (2001)}
\usage{
\method{compute_BIC_MDS}{list}(x, min_dim = 1, max_dim = 5)
}
\arguments{
\item{min_dim}{minimum number of dimensions to test (default = 1)}

\item{max_dim}{maximum number of dimensions to test (default = 5)}

\item{d}{must a \emph{list} of individual symmetric normalized pair-wise distance matrices}
}
\value{
data.frame with the number of dimensions \code{dim}, the stress \code{Stress}, the BIC for a specific number of dimensions \code{BIC}, the residual sum-of-squares \code{RSS} of the used solution, the variance of the distance values accounted for by the solution \code{VAF},  the number of parameters \code{P} according to Lee (2001), and the used precision estiamte \code{s}
}
\description{
Computes the BIC for a specific number of dimensions of the MDS space according to Lee (2001) based on
a list of indidivdual normalized symmetric distance matrices
}
\examples{

\dontrun{

compute_BIC_MDS(d,min_dim = 1, max_dim = 5)

}


}
\author{
David Izydorczyk
}
