% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_BIC_MDS_matrix.R
\name{compute_BIC_MDS.matrix}
\alias{compute_BIC_MDS.matrix}
\title{Compute BIC for MDS according to Lee (2001)}
\usage{
\method{compute_BIC_MDS}{matrix}(x, min_dim = 1, max_dim = 5, s = 0.1)
}
\arguments{
\item{min_dim}{minimum number of dimensions to test (default = 1)}

\item{max_dim}{maximum number of dimensions to test (default = 5)}

\item{s}{Assumed level of data precision. Has to be provided when d is an averaged matrix (default = 0.1). Otherwise s is computed from the individual level data.}

\item{d}{must a symmetric \emph{matrix} of averaged normalized distances}
}
\value{
data.frame with the number of dimensions \code{dim}, the stress \code{Stress}, the BIC for a specific number of dimensions \code{BIC}, the residual sum-of-squares \code{RSS} of the used solution, the variance of the distance values accounted for by the solution \code{VAF},  the number of parameters \code{P} according to Lee (2001), and the used precision estiamte \code{s}
}
\description{
Computes the BIC for a specific number of dimensions of the MDS space according to Lee (2001) based on
a symmetric  \emph{matrix} of averaged normalized distances
}
\examples{

\dontrun{

compute_BIC_MDS(d,min_dim = 1, max_dim = 5, s = 0.1)

}


}
\author{
David Izydorczyk
}
