% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_BIC_MDS.R
\name{compute_BIC_MDS}
\alias{compute_BIC_MDS}
\title{Compute BIC for MDS according to Lee (2001)}
\usage{
compute_BIC_MDS(x, ...)
}
\arguments{
\item{x}{must be either a symmetric  \emph{matrix} of averaged normalized distances or a \emph{list} of individual symmetric normalized pair-wise distance matrices}

\item{min_dim}{minimum number of dimensions to test (default = 1)}

\item{max_dim}{maximum number of dimensions to test (default = 5)}

\item{s}{Assumed level of data precision. Has to be provided when d is an averaged matrix (default = 0.1). Otherwise s is computed from the individual level data.}
}
\value{
data.frame with the number of dimensions \code{dim}, the stress \code{Stress}, the BIC for a specific number of dimensions \code{BIC}, the residual sum-of-squares \code{RSS} of the used solution, the variance of the distance values accounted for by the solution \code{VAF},  the number of parameters \code{P} according to Lee (2001), and the used precision estiamte \code{s}
}
\description{
\code{compute_BIC_MDS} is a generic function that computes the BIC for a specific number of dimensions of the
MDS space according to Lee (2001). The function works on average or individual data, determined by the
\code{\link{class}} of the first argument (i.e., \code{matrix} for average data or \code{list} for indidivdual data)
}
\examples{

\dontrun{

compute_BIC_MDS(d,min_dim = 1, max_dim = 5)

}


}
\seealso{
Other compute_BIC_MDS: 
\code{\link{compute_BIC_MDS.list}()},
\code{\link{compute_BIC_MDS.matrix}()}
}
\author{
David Izydorczyk
}
\concept{compute_BIC_MDS}
