% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dists.R
\name{compute_dists}
\alias{compute_dists}
\title{Compute pairwise distances between two vectors}
\usage{
compute_dists(mat, r = 2, diag = F, upper = F)
}
\arguments{
\item{mat}{matrix with coordinates of stimuli (i.e., the position in multidimensional space), with dimensions in the columns and stimuli in the rows}

\item{r}{r parameter of minkowski space (default = 2)}

\item{diag}{return diagonal matrix values (defualt = FALSE)}

\item{upper}{return upper matrix half (default = FALSE)}
}
\value{
Pairwise distances. By defaul the diagonal and upper triangle of the matrix are \code{NA}
}
\description{
\code{compute_dists} is a wrapper around \code{pair_dist_cpp} to compute pairwise distances between two vectors.
}
\examples{
\dontrun{

compute_dists(matrix(rnorm(10),nrow=5),r = 2)

}

}
\author{
David Izydorczyk
}
