% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_MDS.R
\name{cross_validation_MDS}
\alias{cross_validation_MDS}
\title{Cross-Validation for MDS}
\usage{
cross_validation_MDS(
  mat,
  reps = 100,
  max_dim = 2,
  NA_prob = 0.2,
  verbose = FALSE,
  metric = "r",
  parallel = TRUE,
  n_cores = 4
)
}
\arguments{
\item{mat}{pairwise distance matrix}

\item{reps}{number of repetitions (default = 100)}

\item{max_dim}{maximum number of dimensions to test (default = 2)}

\item{NA_prob}{probability of turning an entry of \code{mat} to \code{NA} (default = 0.2)}

\item{verbose}{prints progress (default = FALSE)}

\item{metric}{CV-metric to use. Either r (correlation) or RMSE (root-mean-squared-error)}

\item{parallel}{run the main computation in parallel}

\item{n_cores}{if parallel == TRUE, how many cores should be used when registering the cluster}
}
\value{
Matrix with number of dimensions \code{dim} and average RMSE \code{m_RMSE} over all \code{reps}.
}
\description{
\code{cross_validation_MDS} runs a cross-validation procedure to find the best fitting number of dimensions of the MDS space
}
\examples{

\dontrun{

cross_validation_MDS(mat,max_dim = 5)

}


}
\author{
David Izydorczyk
}
