% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_mat.R
\name{fill_mat}
\alias{fill_mat}
\title{Fill distance matrix}
\usage{
fill_mat(lower_mat, diag_val = 0)
}
\arguments{
\item{lower_mat}{matrix with only lower triangle entries (rest \code{NA})}

\item{diag_val}{value for diagonal (default = 0)}
}
\value{
full symmetric distance matrix
}
\description{
\code{fill_mat} transforms distance matrix with only lower triangle entries (rest \code{NA}) to full matrix with 0 on the diagonals
}
\examples{

\dontrun{

fill_mat(lower_dist_mat)

}


}
\author{
David Izydorczyk
}
