% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_MDS.R
\name{gen_data_MDS}
\alias{gen_data_MDS}
\title{Generate MDS data (i.e., true underlying dimensions and true pairwise distances)}
\usage{
gen_data_MDS(
  ndims = 2,
  n = 16,
  min = -1,
  max = 1,
  r = 2,
  diag = F,
  upper = F,
  return_latent = T
)
}
\arguments{
\item{ndims}{number of latent dimensions/cues (default = 2)}

\item{n}{number of itmes/rows (default = 16)}

\item{min}{minimum possible cue value for the uniform distribution (default = -1)}

\item{max}{maximum possible cue value for the uniform distribution (default = 1)}

\item{r}{r parameter of minkowski space (default = 2)}

\item{diag}{return diagonal matrix values (defualt = FALSE)}

\item{upper}{return upper matrix half (default = FALSE)}

\item{return_latent}{should the latent matrix of underlying cues and cue values be returned? (default = TRUE)}
}
\value{
matrix of pairwise-distances between objects (if return_latent = FALSE) or a list with the matrix of pairwise distances (\code{dist_mat}) and the underlying MDS space (\code{latent})
}
\description{
\code{gen_data_MDS} generates a matrix with thetrue underlying dimensions and the corresponding true pairwise distances
}
\examples{

gen_data_MDS()

}
\author{
David Izydorczyk
}
