% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_CV_MDS.R
\name{loo_CV_MDS}
\alias{loo_CV_MDS}
\title{Leave-One-Out Cross-Validation for MDS}
\usage{
loo_CV_MDS(mat, max_dim = 2, verbose = FALSE, parallel = TRUE, n_cores = 4)
}
\arguments{
\item{mat}{pairwise distance matrix}

\item{max_dim}{maximum number of dimensions to test (default = 2)}

\item{verbose}{prints progress (default = FALSE)}

\item{parallel}{run the main computation in parallel}

\item{n_cores}{if parallel == TRUE, how many cores should be used when registering the cluster}
}
\value{
Matrix with number of dimensions \code{dim} and R-squared, RMSE and MSE between the true and predicted distances in each leave-one-out trial
}
\description{
\code{loo_CV_MDS} runs a leave-one-out (loo) cross-validation procedure to find the best fitting number of dimensions of the MDS space
}
\examples{

\dontrun{

loo_CV_MDS(mat,max_dim = 5,verbose = TRUE)

}


}
\author{
David Izydorczyk
}
