% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lower_mat.R
\name{make_lower_mat}
\alias{make_lower_mat}
\title{Make a lower traingular part of distance matrix}
\usage{
make_lower_mat(d_mat)
}
\arguments{
\item{d_mat}{symmetric matrix}
}
\value{
lower triangular matrix, with rest \code{NA}
}
\description{
\code{make_lower_mat} makes a lower triangular matrix out of a symmetric distance matrix
}
\examples{

\dontrun{

make_lower_mat(d_mat)

}


}
\author{
David Izydorczyk
}
