% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_overview.R
\docType{package}
\name{MDShelper}
\alias{MDShelper-package}
\alias{MDShelper}
\title{MDShelper: MDS Helper Functions}
\description{
This package contains a bundle of convenience and helper functions for conducting multi-dimensional scaling analysis.
}
\section{Functions}{

\itemize{
\item \code{\link{gen_lat_dim}}:
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dizyd/MDShelper}
  \item Report bugs at \url{https://github.com/dizyd/MDShelper/issues}
}

}
\author{
\strong{Maintainer}: David Izydorczyk \email{izydorczyk@uni-mannheim.de} (\href{https://orcid.org/0000-0002-8792-0795}{ORCID})

}
