% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC_MDS.R
\name{BIC_MDS}
\alias{BIC_MDS}
\title{Compute BIC for MDS according to Lee (2001)}
\usage{
BIC_MDS(s, rss, n_dims, n_items)
}
\arguments{
\item{s}{sample estimate of standard deviation}

\item{rss}{residual sum-of-squares of current solution}

\item{n_dims}{number of dimensions number}

\item{n_items}{number of items}
}
\value{
BIC for a specific number of dimensions.
}
\description{
\code{BIC_MDS} computes the BIC for a specific number of dimensions of the MDS space according to Lee (2001)
}
\examples{

\dontrun{

BIC_MDS()

}


}
\author{
David Izydorczyk
}
