% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_BIC_MDS.R
\name{compute_BIC_MDS}
\alias{compute_BIC_MDS}
\title{Compute BIC for MDS according to Lee (2001)}
\usage{
compute_BIC_MDS(d_list, max_dim = 5)
}
\arguments{
\item{d_list}{list of pair-wise distance matrices (upper + diag should be NA)}

\item{max_dim}{maximum number of dimensions to test (default = 5)}
}
\value{
Matrix with the number of dimensions \code{dim}, the stress \code{stress}, the BIC for a specific number of dimensions \code{BIC}, the residual sum-of-squares \code{rss}, and the number of parameters \code{P} according to Lee (2001), and the estimated precision estiamte \code{s}
}
\description{
\code{BIC_MDS} computes the BIC for a specific number of dimensions of the MDS space according to Lee (2001)
}
\examples{

\dontrun{

compute_BIC_MDS(list_of_distances)

}


}
\author{
David Izydorczyk
}
