% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_overview.R
\docType{package}
\name{MDShelper}
\alias{MDShelper}
\title{MDShelper: MDS Helper Functions}
\description{
This package contains a bundle of convenience and helper functions for conducting multi-dimensional scaling analysis.
}
\section{Functions}{

\itemize{
\item \code{\link{gen_lat_dim}}:
}
}

